# api/openapi.yaml
openapi: "3.1.0"
info:
  title: jobifai API
  description: AI-powered job application automation, REST API
  version: "1.0.0"
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/bot/review/{job_id}/approve:
    post:
      operationId: reviewApprove
      summary: Approve a pending application, bot will submit it
      tags: [bot]
      parameters:
        - $ref: "#/components/parameters/JobIdPath"
      responses:
        "200":
          $ref: "#/components/responses/OkResponse"
        "404":
          description: No pending review for this job_id
  /api/bot/review/{job_id}/reject:
    post:
      operationId: reviewReject
      summary: Reject a pending application, bot will skip it
      tags: [bot]
      parameters:
        - $ref: "#/components/parameters/JobIdPath"
      responses:
        "200":
          $ref: "#/components/responses/OkResponse"
        "404":
          description: No pending review for this job_id

  # ── Job history ────────────────────────────────────────────────────────
...
components:
  parameters:
...
    JobIdPath:
      name: job_id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    OkResponse:
      description: Success
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/OkMessage"
  schemas:
...
    OkMessage:
      type: object
      properties:
        message:
          type: string
      required: [message]
...